update to 11.4.2 (5469)

+++ TMessagesProj/src/main/java/org/telegram/ui/Components/Text.java
@@ -93,8 +93,8 @@ public class Text {
         paint.setColor(color);
     }
 
-    private int ellipsizeWidth = -1;
-    public Text ellipsize(int width) {
+    private float ellipsizeWidth = -1;
+    public Text ellipsize(float width) {
         ellipsizeWidth = width;
         return this;
     }
